fix values
