+++ packages-kintone-customize/app-229-contracts/src/view/indexShow/renderCumulative/sections/filter/SelectYear.tsx
@@ -2,11 +2,13 @@ import { FormControl, InputLabel, MenuItem, Select } from '@mui/material';
 import { useMemo } from 'react';
 import { useTypedFormContext } from '../../hooks/useTypedRHF';
 import { Controller } from 'react-hook-form';
+import { getLatestMonths } from '../../../../../helpers/getLatestMonths';
 
 
 export const SelectYear = () => {
   const {
     control,
+    setValue,
   } = useTypedFormContext();
 
   const years = useMemo(
@@ -47,7 +49,10 @@ export const SelectYear = () => {
               id="selectYear"
               value={value}
               label="年度"
-              onChange={(e) => onChange(e.target.value)}
+              onChange={(e) => {
+                onChange(e.target.value);
+                setValue('months', getLatestMonths(+e.target.value, 3));
+              }}
               {...otherField}
             >
               {years.map((year) => (
