+++ packages-kintone-customize/app-229-contracts/src/view/indexShow/renderCumulative/sections/results/fiscalMonths/FiscalMonths.tsx
@@ -20,6 +20,8 @@ export const FiscalMonths = ({
     data,
   } = fiscalYearQuery;
 
+  console.log(data);
+
   return (
     <>
       {selectedMonths.map((month) => {
